requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.30.2
azure-storage-blob==12.21.0
azure-storage-file-share==12.17.0
bandit==1.7.8
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-utils==0.6.0
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-docstrings-complete==1.3.0
flake8-functions==0.0.8
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mr-proper==0.0.7
mypy==1.10.0
mypy-extensions==1.0.0
orjson==3.10.5
packaging==24.1
pathspec==0.12.1
pbr==6.0.0
pep8-naming==0.14.1
platformdirs==4.2.2
psutil==5.9.8
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
PyMySQL==1.1.1
pyodbc==5.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.30
sqlalchemy-stubs==0.4
sqlalchemy2-stubs==0.0.2a38
starlette==0.37.2
stdlib-list==0.10.0
stevedore==5.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
